% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{CV_delta}
\alias{CV_delta}
\title{Cross validation to select \eqn{\delta}}
\usage{
CV_delta(X, deltaGrids, diagonal, se_est, merge)
}
\arguments{
\item{X}{A \eqn{n} by \eqn{p} data matrix.}

\item{deltaGrids}{A vector of numerical constants.}

\item{diagonal}{Logical. If TRUE, the covariance matrix of \eqn{Z} is
diagonal; else FALSE.}

\item{se_est}{The vector of the standard deviations of \eqn{p} features.}

\item{merge}{Logical. If TRUE, take the union of all candidate pure variables;
otherwise, take the intersection.}
}
\value{
A numeric constant. The selected optimal \eqn{\delta}.
}
\description{
Cross validation for choosing the tuning parameter \eqn{\delta}.
  For each value of \code{deltaGrids}, first split the data into two parts
  and calculate \eqn{I}, \eqn{A_I} and \eqn{Cov(Z)}. Then calculate the fit
  \eqn{A_I Cov(Z) A_I'} to find the value which minimizes the loss criterion
  \deqn{||\Sigma - A_I Cov(Z) A_I'||_{F-off}/(|I|(|I|-1))}.
}
