% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstOmega.R
\name{estOmega}
\alias{estOmega}
\title{Estimation of the precision matrix}
\usage{
estOmega(lbd, C)
}
\arguments{
\item{lbd}{A numeric constant.}

\item{C}{A \eqn{K} by \eqn{K} matrix.}
}
\value{
A \eqn{K} by \eqn{K} matrix.
}
\description{
\code{estOmega} estimates the inverse matrix of \code{C} via regularizations.
}
\details{
For a given matrix \code{C}, \code{estOmega} finds its inverse
  by solving the following linear program
  \deqn{\min_{\Omega} ||\Omega||_{\infty,1}}
  subject to
  \deqn{||C \Omega - I|| \le \lambda.}
  The above LP is solved by decoupling into several linear programs, each of
  which solves one row of \eqn{\Omega}.
}
