% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{KfoldCV_delta}
\alias{KfoldCV_delta}
\title{Cross-validation for selecting \eqn{delta}}
\usage{
KfoldCV_delta(
  X,
  delta = NULL,
  ndelta = 50,
  q = 2,
  exact = FALSE,
  nfolds = 10,
  max_pure = NULL
)
}
\arguments{
\item{X}{A \eqn{n} by \eqn{p} data matrix.}

\item{delta}{The grid of leading constant of \eqn{\delta}.}

\item{ndelta}{Integer. The length of the grid of \code{delta}.}

\item{q}{Either \code{2} or \code{Inf} to specify the type of score.}

\item{exact}{Logical. Only active for compute the \code{Inf} score.
If TRUE, compute the \code{Inf} score exactly via solving a linear program.
Otherwise, use approximation to compute \code{Inf} score.}

\item{nfolds}{The number of folds. Default is 10.}

\item{max_pure}{A numeric value between (0, 1] specifying the maximal
proportion of pure variables. Default is NULL. When not specified,
\code{max_pure} = 1 if \eqn{n > p}, \code{max_pure} = 0.8 otherwise.}
}
\value{
A list of objects including: \itemize{
  \item\code{foldid} The indices of observations used for cv.
  \item\code{delta_min} The value of \code{delta} that has the minimum cv
    error.
  \item\code{delta_1se} The smallest and largest value of \code{delta} such
    that the cv errors are wihtin one standard error of the minimum cv error.
  \item\code{delta} The used \code{delta} sequence.
  \item\code{cv_mean} The averaged cv errors.
  \item\code{cv_sd} The standard errors of cv errors.
  \item\code{est_pure} A list including: \itemize{
    \item\code{K} The cardinality of parallel rows.
    \item\code{I} The index of parallel rows.
    \item\code{I_part} The partition of parallel rows.
  }
  \item\code{score} The score matrix.
  \item\code{moments} The crossproduct matrix \eqn{R'R}.
}
}
\description{
\code{KfoldCV_delta} uses \code{nfolds} cross-validation to
  select \eqn{delta}.
}
